deixando de dar watch só se já estiver watching

+++ src/main/java/br/com/caelum/brutal/dao/WatcherDAO.java
@@ -56,13 +56,15 @@ public class WatcherDAO {
 		if(!alreadyWatching(watcher)) {
 			add(watcher);
 		} else {
-			remove(watcher);
+			removeIfWatching(watcher);
 		}
 	}
 
-	public void remove(Watcher watcher) {
+	public void removeIfWatching(Watcher watcher) {
 		Watcher managed = findByQuestionAndUser(watcher.getWatchedQuestion(), watcher.getWatcher());
-		session.delete(managed);
+		if (managed != null) {
+			session.delete(managed);
+		}
 	}
 
 	private boolean alreadyWatching(Watcher watcher) {
